+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.0.90 (2026-03-28)
 
 * [Use intern! for enum getattr on dump path](https://github.com/anna-money/marshmallow-recipe/pull/286)
+* [Optimize _get_container cache lookup from 3 probes to 1](https://github.com/anna-money/marshmallow-recipe/pull/287)
 
 ## v0.0.89 (2026-03-27)